third_party/WebKit/LayoutTests/svg/filters/feBlend-invalid-mode-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: Type error
PASS: Invalid blend modes do not trigger a crash. 
